Introduce Fibonacci with big number
fib(93) is error because overflow
introduce big number by a bn struct bn
typedef struct _bn {
    unsigned int *number;
    unsigned int size;
    int sign;
    int capacity;
} bn;

fib(1000) could be calculated correctly,
but not fully exploit 64 bits processor.

Potential impacts:

The time consumed of calculating fib(n), n < 93
will be longer, because all numbers are
represented using struct _bn(even if these
numbers do not require the use of struct _bn).

Arithmetic operations are performed specifically
for struct _bn rather than the original unsigned
int addition, subtraction, and multiplication
operations, this will consume more longer time.
------------------------------------------
Note that there is still a limit on the
size of numbers that can be processed even
with big number.

kmalloc has limits on the amount of memory
allocated, and allocating beyond a certain
size can result in errors. The limit depends
on VMM. Suppose that the limit is 4 MB,
according to Binet's formula, the size of
fib(5761680) is 4 MB. If n > 5761680, the
error will be caused due to malloc fails.
-------------------------------------------